Добавил логику, что при выборе пункта назначения - меняем описание
